Windows-Printer-Client/requirements.txt
python-socketio[client]==5.10.0
python-dotenv==1.0.0
pywin32>=307
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pystray>=0.19.0
weasyprint>=60.0
reportlab>=4.0.0
